% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_add_equation.R
\name{add_equation_single}
\alias{add_equation_single}
\title{Add single equation to the model}
\usage{
add_equation_single(model, equation, hidden = FALSE, desc = "")
}
\description{
helper for add_equation()
}
