% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_prepare_model.R
\name{add_lag_info}
\alias{add_lag_info}
\title{Find dependencies and order the equations}
\usage{
add_lag_info(x)
}
\arguments{
\item{x}{A vector to modify}
}
\description{
Find dependencies and order the equations
}
