% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_add_scenario.R
\name{add_scenario}
\alias{add_scenario}
\title{Make initial matrix  row for scenario and prepare shocks}
\usage{
add_scenario(
  model,
  name = "expansion",
  origin = "baseline",
  origin_period = 1,
  shocks = shock
)
}
\arguments{
\item{model}{a SFC model}

\item{name}{scenario name}

\item{origin}{origin scenario which from want to simulate}

\item{origin_period}{origin period in origin scenario which from want to simulate.
Defaults 1.}

\item{shocks}{tibble making using \code{create_shock()} and \code{add_shock()}}
}
\description{
Make initial matrix  row for scenario and prepare shocks
}
