% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shock.R
\name{add_shock}
\alias{add_shock}
\title{Define chenge of exogenous variable and time for shocking selected scenario}
\usage{
add_shock(shock, equation = "", desc = "", start = NA, end = NA)
}
\arguments{
\item{shock}{a tibble object after \code{create_shock}}

\item{equation}{equation  string in format: 'x = 2'}

\item{desc}{variable description}

\item{start}{Defaults NA.}

\item{end}{Defaults NA.}
}
\value{
an updated shock object containing information
}
\description{
Define chenge of exogenous variable and time for shocking selected scenario
}
