% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_create_add_shock.R
\name{add_shock_vector}
\alias{add_shock_vector}
\title{Add vector as a shock to shock object}
\usage{
add_shock_vector(shock, variable, values, desc = "", start = 1)
}
\arguments{
\item{shock}{tibble from \code{create_shock()}}

\item{variable}{variable}

\item{values}{values}

\item{desc}{string variable description}

\item{start}{numeric period number for the shock to take place, defaults to NA}

\item{end}{numeric period number for the shock to take place, defaults to NA}
}
\value{
updated shock object containing added shock
}
\description{
Add vector as a shock to shock object
}
