% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_add_variables.R
\name{add_variable}
\alias{add_variable}
\title{Adds variables to the model}
\usage{
add_variable(model, name = "", init = 1e-15, desc = "")
}
\arguments{
\item{model}{a SFC object}

\item{name}{variable name}

\item{init}{initial value. Defaults to 1.00e-15.}

\item{desc}{variable description}
}
\value{
an updated SFC object containing variables
}
\description{
Adds variables to the model
}
