% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_add_variable.R
\name{add_variable_single}
\alias{add_variable_single}
\title{Add single variable to the model}
\usage{
add_variable_single(model, name, init = NA, desc = "")
}
\arguments{
\item{model}{SFC model object}

\item{name}{string name for added variable}

\item{init}{numeric initial value, defaults to 1e-05}

\item{desc}{string variable description}
}
\description{
helper for add_variable()
}
