% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_change_init.R
\name{change_init}
\alias{change_init}
\title{Change initial value of a variable}
\usage{
change_init(model, name, value)
}
\arguments{
\item{model}{SFC model object}

\item{name}{string variable name}

\item{value}{numeric value that will replace existing initial value}
}
\value{
updated SFC model object with new variable initial value
}
\description{
Change initial value of a variable
}
