% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_01_validate_model_input.R
\name{check_parentheses_balance}
\alias{check_parentheses_balance}
\title{Verify if parentheses in the equation are properly balanced}
\usage{
check_parentheses_balance(eq)
}
\arguments{
\item{eq}{A character string representing an equation.}
}
\value{
A logical value indicating whether the parentheses are balanced (`TRUE`) or not (`FALSE`).
}
\description{
This function checks if all parentheses in a given equation string are correctly
paired and balanced. If there are no parentheses, the equation is considered
trivially balanced. It returns `TRUE` if the parentheses are balanced and `FALSE` otherwise.
}
\keyword{internal}
