% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_create_sensitivity.R
\name{create_sensitivity}
\alias{create_sensitivity}
\title{Create model with sensitivity scenarios}
\usage{
create_sensitivity(model_pass, variable, lower = 0, upper = 1, step = 0.1)
}
\arguments{
\item{model_pass}{SFC model object that will be used as a baseline for sensitivity calculation}

\item{variable}{string name of the variable that will be used}

\item{lower}{numeric lower bound value of the variable}

\item{upper}{numeric upper bound value of the variable}

\item{step}{numeric step between upper and lower bounds for the variable to take value}
}
\value{
SFC model object with sensitivity scenarios
}
\description{
Create model with sensitivity scenarios
}
