% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_shock.R
\name{create_shock}
\alias{create_shock}
\title{Constructor shock tibble used in \code{add_scenario()}.}
\usage{
create_shock()
}
\value{
Tibble object
}
\description{
Constructor shock tibble used in \code{add_scenario()}.
}
