% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_simulate_scenario.R
\name{d}
\alias{d}
\title{Calculate 1 order lag difference of a variable in model}
\usage{
d(x)
}
\arguments{
\item{x}{variable name}
}
\value{
difference
}
\description{
Calculate 1 order lag difference of a variable in model
}
\details{
this is a special function to be used exclusively in model equation strings e.g. "x = d(y) + z"
}
