% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_prepare.R
\name{find_adjacency}
\alias{find_adjacency}
\title{Find adjacency matrix for system of equations}
\usage{
find_adjacency(equations)
}
\arguments{
\item{equations}{system of equations already time stamped}
}
\value{
adjacency matrix
}
\description{
Find adjacency matrix for system of equations
}
\author{
João Macalós
}
