% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_prepare.R
\name{find_blocks}
\alias{find_blocks}
\title{Find blocks of independent equations (using \code{igraph} functions)}
\usage{
find_blocks(adj)
}
\arguments{
\item{adj}{adjacency matrix}
}
\description{
Find blocks of independent equations (using \code{igraph} functions)
}
\author{
João Macalós
}
