% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_plot_cycles.R
\name{plot_cycles}
\alias{plot_cycles}
\title{Network plot of the model}
\usage{
plot_cycles(model, save_file = NULL)
}
\arguments{
\item{model}{SFC model object created with \code{\link{create_model}}}

\item{save_file}{name and path to save the plot as html file}
}
\value{
visNetwork object
}
\description{
Network plot of the model
}
\details{
This function creates a representation of a model as a directed graph.
Additionally it shows cycles in the model including these with lagged variables.
Graph can be saved as html file.
}
\examples{
model <- godley::create_model(name = "SFC model", template = "BMW")
plot_cycles(model)

}
