% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_plot.R
\name{plot_series}
\alias{plot_series}
\title{Plotting series comparison of a given model' scenario.}
\usage{
plot_series(
  model,
  scenario = "expansion",
  from = 1,
  to = NA,
  expressions = list("Y", "G_s / alpha1")
)
}
\arguments{
\item{model}{a SFC model object, used mostly for enriching a plot}

\item{scenario}{a name of scenario from which take variables values}

\item{from}{first iteration to plot}

\item{to}{last iteration to plot}

\item{expressions}{a vector (or single character string) containing information
which variable/s expressions to plot}
}
\value{
plotly plot object ready to show / save
}
\description{
Maximum number of series is 5 (to properly choose color from a vector)
}
