% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_plot_simulation_g.R
\name{plot_simulation_g}
\alias{plot_simulation_g}
\title{Plot simulations of multiple variables in multiple scenarios}
\usage{
plot_simulation_g(
  model,
  scenario = "baseline",
  take_all = FALSE,
  from = NA,
  to = NA,
  expressions = "Y",
  colors = NULL,
  max.overlaps = 10
)
}
\arguments{
\item{model}{SFC model object}

\item{scenario}{vector of strings or single string name of scenario(s) from which take variables values, defaults to 'baseline'}

\item{take_all}{logical indicating whether all scenarios containing the given scenario name string(s) should be used, defaults to FALSE}

\item{from}{numeric period number from which the plot should start, defaults to maximum value}

\item{to}{numeric period number on which the plot should end, defaults to minimum value}

\item{expressions}{vector of strings or single string name of variable(s) expression(s) to plot, defaults to 'Y'. Accepts basic math formulas: `+`, `-`, `*`, `/`}

\item{colors}{vector of strings or single string with desired color names}

\item{max.overlaps}{sets max.overlaps as in ggrepel::geom_text_repel()}
}
\value{
ggplot2 plot
}
\description{
Returns ggplot2 plots which could be further tweaked.
}
\details{
This functions is created to help preparing publication ready plots. As any normal ggplot2 plot it can be
modified using any function from the ggplot2 universe.

`expressions` allows to create own variables like "W/Y"

To remove unnecessary legends use + guides(colour = "none", linetype = "none") after the function call.
}
\seealso{
[plot_simulation()] for interactive plots best for model testing.
}
