% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_02_newton.R
\name{prep_broyden}
\alias{prep_broyden}
\title{Prep equations for  Newton solvers}
\usage{
prep_broyden(.block)
}
\arguments{
\item{.block}{Blocks of equations}
}
\value{
blocks
}
\description{
Prep equations for  Newton solvers
}
\author{
João Macalós
}
