% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_prepare.R
\name{prep_equations}
\alias{prep_equations}
\title{Re-write the equations with the correct matrix syntax that will be used to evaluate the expressions inside the Gauss Seidel/Newton algorithm}
\usage{
prep_equations(equations_sep, variables_exo)
}
\arguments{
\item{equations_sep}{tibble of equations after passing \code{validate_model_input()} function}

\item{variables_exo}{tibble of exogenous values after passing \code{validate_model_input()} function}
}
\value{
tibble of equations with the correct matrix syntax
}
\description{
Re-write the equations with the correct matrix syntax that will be used to evaluate the expressions inside the Gauss Seidel/Newton algorithm
}
