% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_prepare_model.R
\name{prepare}
\alias{prepare}
\title{Making initial matrix row for baseline scenario and prepare equations for \code{simulate()}}
\usage{
prepare(model)
}
\arguments{
\item{model}{a SFC object}
}
\value{
indicator that model is prepared
}
\description{
Making initial matrix row for baseline scenario and prepare equations for \code{simulate()}
}
