% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_00_prepare_scenario_matrix.R
\name{prepare_scenario_matrix}
\alias{prepare_scenario_matrix}
\title{Making initial matrix row for scenario specified by \code{create_shock()} and
\code{add_scenario()}.}
\usage{
prepare_scenario_matrix(m, shock)
}
\arguments{
\item{shock}{Tibble created by \code{create_shock()}}

\item{model}{a SFC object}
}
\value{
Matrix with initial matrix ready to \code{simulate()}
}
\description{
Making initial matrix row for scenario specified by \code{create_shock()} and
\code{add_scenario()}.
}
