% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_01_gauss_seidel.R
\name{run_gauss_seidel}
\alias{run_gauss_seidel}
\title{Gauss Seidel algorithm}
\usage{
run_gauss_seidel(m, calls, periods, max_iter, tol)
}
\arguments{
\item{m}{The initialized matrix obtained with \code{prepare()} or
\code{prepare_scenario_matrix()}}

\item{calls}{Prepared equations with \code{prepare()}.}

\item{periods}{Total number of rows (periods) in the model.}

\item{max_iter}{Maximum number of iterations allowed per block per period.}

\item{tol}{Tolerance accepted to determine convergence.}
}
\value{
simulated scanario matrix
}
\description{
Gauss Seidel algorithm
}
\details{
This algorithm simulates the model by recursion by using
nested for loops. At each round of iteration, the values calculated
are compared to the previous values. If the difference is below
a tolerance value set by the user, the round of calculations have converged
and the algorithm jump to the next block of equations.

The algorithm modifies a matrix in place to optimize its performance.
}
\author{
João Macalós
}
