% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_02_newton.R
\name{run_newton}
\alias{run_newton}
\title{Newton-Raphson solver implemented with \code{rootSolve::multiroot()}}
\usage{
run_newton(m, calls, periods, max_iter, tol, ...)
}
\arguments{
\item{m}{The initialized matrix obtained with code{prepare()} or
\code{prepare_scenario_matrix()}}

\item{calls}{Prepared equations with \code{prepare()}.}

\item{periods}{Total number of rows (periods) in the model.}

\item{max_iter}{Maximum number of iterations allowed per block per period.}

\item{tol}{Tolerance accepted to determine convergence.}
}
\value{
simulated scanario matrix
}
\description{
Newton-Raphson solver implemented with \code{rootSolve::multiroot()}
}
\details{
This function implements the Newton-Raphson method to solve the cyclical
blocks of equations. It relies on the \code{multiroot()} function from \code{rootSolve}.
}
\author{
João Macalós
}
