% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate the scenario of a stock-flow consistent model}
\usage{
simulate(
  model,
  scenario = "baseline",
  max_iter = 350,
  periods = 100,
  hidden_tol = 0.1,
  tol = 1e-08,
  method = "Gauss"
)
}
\arguments{
\item{model}{SFC model}

\item{scenario}{Name of scenario want to simulate}

\item{max_iter}{Maximum iterations allowed per period.}

\item{periods}{Total number of rows (periods) in the model.}

\item{tol}{Tolerance accepted to determine convergence.}

\item{method}{The method to use to find a solution. Defaults to "Gauss".}

\item{.hidden_tol}{Error tolerance to accept the equality of the hidden equation. Defaults to 0.1.}
}
\value{
Updated model containing scenario simulate result.
}
\description{
Simulate the scenario of a stock-flow consistent model
}
