% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_01_validate_user_data.R
\name{validate_data}
\alias{validate_data}
\title{'Validates' a model (checks conditions and verify information introduced by user )}
\usage{
validate_data(model)
}
\arguments{
\item{model}{a SFC object}
}
\value{
an updated equation and external values
}
\description{
Checking: initial conditions, defined / not defined variables existence
}
