% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_add_equations.R
\name{add_equation}
\alias{add_equation}
\title{Adds an equation to the model}
\usage{
add_equation(model, equation = "", desc = "", hidden = FALSE)
}
\arguments{
\item{model}{a SFC object}

\item{equation}{equation string in format: 'x = y + z + w + 2'}

\item{desc}{equation description, e.g. consumption flow}

\item{hidden}{logical, indicates if equation is hidden}
}
\value{
an updated SFC object containing equations
}
\description{
Adds an equation to the model
}
