% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_add_equation.R
\name{add_equation}
\alias{add_equation}
\title{Add equations to the model}
\usage{
add_equation(model, ...)
}
\arguments{
\item{model}{SFC model object}

\item{...}{additional arguments passed to the function.}
}
\value{
updated SFC model object containing added equation
}
\description{
Add equations to the model
}
