% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_add_equation.R
\name{add_equation_single}
\alias{add_equation_single}
\title{Add single equation to the model}
\usage{
add_equation_single(model, equation, hidden = FALSE, desc = "")
}
\arguments{
\item{model}{SFC model object}

\item{equation}{string equation in format: 'x = y + z - a * b + (c + d) / e + f[-1]'}

\item{hidden}{logical, indicates if equation should be written as hidden, defaults to FALSE}

\item{desc}{string equation description}
}
\description{
helper for add_equation()
}
