% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_add_scenario.R
\name{add_scenario}
\alias{add_scenario}
\title{Add scenario to the model}
\usage{
add_scenario(
  model,
  name = "expansion",
  origin = "baseline",
  origin_start = NA,
  origin_end = NA,
  shock
)
}
\arguments{
\item{model}{SFC model object}

\item{name}{string name of scenario, defaults to 'expansion'}

\item{origin}{string name of origin scenario, from which the new scenario will be created, defaults to 'baseline'}

\item{origin_start}{numeric period number from origin scenario from which the new scenario will begin}

\item{origin_end}{numeric period number from origin scenario on which the new scenario will end}

\item{shock}{shock object from \code{create_shock()} and \code{add_shock()}}
}
\value{
updated SFC model object containing added scenario
}
\description{
Add scenario to the model
}
