% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_create_add_shock.R
\name{add_shock}
\alias{add_shock}
\title{Add shock to shock object}
\usage{
add_shock(shock, equation, desc = "", start = NA, end = NA)
}
\arguments{
\item{shock}{tibble from \code{create_shock()}}

\item{equation}{string equation in format: 'x = 2'}

\item{desc}{string variable description}

\item{start}{numeric period number for the shock to take place, defaults to NA}

\item{end}{numeric period number for the shock to take place, defaults to NA}
}
\value{
updated shock object containing added shock
}
\description{
Add shock to shock object
}
