% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_create_add_shock.R
\name{add_shock}
\alias{add_shock}
\title{Add shock to shock object}
\usage{
add_shock(
  shock,
  variable,
  value = NA,
  rate = NA,
  absolute = NA,
  start = NA,
  end = NA,
  desc = ""
)
}
\arguments{
\item{shock}{tibble from \code{create_shock()}}

\item{variable}{string variable name}

\item{value}{numeric, an explicit value or values for the variable, will be extended with last value}

\item{rate}{numeric, multiplier to influence the original value of the variable}

\item{absolute}{numeric, absolute value to influence the original value of the variable}

\item{start}{numeric or date period number for the shock to take place, defaults to NA}

\item{end}{numeric or date period number for the shock to take place, defaults to NA}

\item{desc}{string variable description}
}
\value{
updated shock object containing added shock
}
\description{
Add shock to shock object
}
