% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_add_variable.R
\name{add_variable}
\alias{add_variable}
\title{Add variable to the model}
\usage{
add_variable(model, name, init = 0, desc = "")
}
\arguments{
\item{model}{SFC model object}

\item{name}{string name for added variable}

\item{init}{numeric initial value, defaults to 1e-15}

\item{desc}{string variable description}
}
\value{
updated SFC model object containing added variable
}
\description{
Add variable to the model
}
