% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_add_variable.R
\name{add_variable}
\alias{add_variable}
\title{Add variables to the model}
\usage{
add_variable(model, ...)
}
\arguments{
\item{model}{SFC model object}

\item{...}{additional arguments passed to the function.}
}
\value{
updated SFC model object containing added variable
}
\description{
Add variables to the model
}
