% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_create_model.R
\name{create_model}
\alias{create_model}
\title{SFC model constructor}
\usage{
create_model(name = "SFC model")
}
\arguments{
\item{name}{name for created SFC model}
}
\value{
SFC object
}
\description{
SFC model constructor
}
