% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_create_model.R
\name{create_model}
\alias{create_model}
\title{Create SFC model object}
\usage{
create_model(name = "SFC model", template)
}
\arguments{
\item{name}{string name for created SFC model object}

\item{template}{string name of model template chosen from: 'SIM', 'PC', 'LP', 'REG', 'OPEN', 'BMW', 'BMWK', 'DIS', 'DISINF', 'SIMEX', 'PCEX'
or user created SFC model object to be used as a template}
}
\value{
SFC model object
}
\description{
Create SFC model object
}
