% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_create_add_shock.R
\name{create_shock}
\alias{create_shock}
\title{Create shock used in \code{add_scenario()}.}
\usage{
create_shock()
}
\value{
shock object
}
\description{
Create shock used in \code{add_scenario()}.
}
