% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_03_run_newton.R
\name{.newton_solver}
\alias{.newton_solver}
\title{Newton solver algorithm}
\usage{
.newton_solver(.x0, .fn, max_iter, tol)
}
\arguments{
\item{.x0}{Vector with initial guess for x.}

\item{.fn}{A function containing the system of equations.}

\item{max_iter}{Maximum number of iterations allowed}

\item{tol}{A numeric value indicating the accepted tolerance to declare convergence.}
}
\description{
Newton solver algorithm
}
\note{
Check https://www.math.usm.edu/lambers/mat419/lecture11.pdf for a quick reference
on the algorithm.
}
\author{
...
}
\keyword{internal}
