% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_prepare_model.R
\name{prep_equations}
\alias{prep_equations}
\title{Re-wrote the equations with the correct matrix syntax that will be used to evaluate
the expressions inside the Gauss Seidel/Newton algorithm}
\usage{
prep_equations(eqs_separated, external_values)
}
\arguments{
\item{eqs_separated}{Tibble of equations after passing  \code{validate_data()} function.}

\item{external_values}{Tibble of exogenous values after passing  \code{validate_data()} function.}
}
\value{
Tibble of equations with the correct matrix syntax
}
\description{
Re-wrote the equations with the correct matrix syntax that will be used to evaluate
the expressions inside the Gauss Seidel/Newton algorithm
}
