% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_prepare.R
\name{prepare}
\alias{prepare}
\title{Make initial matrix row for baseline scenario and prepare equations for \code{simulate_scenario()}}
\usage{
prepare(model, verbose = FALSE)
}
\arguments{
\item{model}{SFC model object}

\item{verbose}{logical to tell if additional model verbose should be displayed}
}
\value{
verified and prepared SFC model object
}
\description{
Make initial matrix row for baseline scenario and prepare equations for \code{simulate_scenario()}
}
