% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_02_run_gauss_seidel.R
\name{restore_equation}
\alias{restore_equation}
\title{Restore equation notation for warning messages}
\usage{
restore_equation(eq_line)
}
\arguments{
\item{eq_line}{A character string representing an equation with internal indexing.}
}
\value{
A character string where the internal indexing notation has been restored to 
a more user-friendly form.
}
\description{
This function is designed to take an internally processed equation string, 
commonly used in model computations (e.g., `m[.i, 'Var']`), and restore it 
to a more original or readable format (e.g., `Var`). It is particularly 
useful for warning messages, where presenting the equation in its original 
form can help the user identify issues more quickly. The function also 
handles lagged variables (e.g., `m[.i - 1, 'Var']` → `Var[-1]`).
}
