% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_04_run_broyden.R
\name{run_broyden}
\alias{run_broyden}
\title{Broyden solver wrapper}
\usage{
run_broyden(m, calls, periods, max_iter, tol, deps, ...)
}
\arguments{
\item{m}{The initialized matrix obtained with \code{.make_matrix()}.}

\item{calls}{Prepared equations with \code{.prep_equations()}.}

\item{periods}{Total number of rows (periods) in the model.}

\item{max_iter}{Maximum number of iterations allowed per block per period.}

\item{tol}{Tolerance accepted to determine convergence.}
}
\description{
Broyden solver wrapper
}
\details{
This function implements the Broyden method to solve the cyclical
blocks of equations.
}
\author{
João Macalós
}
\keyword{internal}
